+++ run-tests.ps1
@@ -1,5 +1,3 @@
-#ps1_sysnative
-
 # Copyright 2014 Cloudbase Solutions Srl
 #
 #    Licensed under the Apache License, Version 2.0 (the "License"); you may