+++ Changelog.md
@@ -1,3 +1,20 @@
+###version 1.00
+  - 21.08.2015
+  - Added ``Carbon::Generate-StrongPassword``
+  - Added ``Juju::Is-Leader``
+  - Added ``Juju::Get-LeaderData``
+  - Added ``Juju::Set-LeaderData``
+  - Added ``Juju::Get-JujuRemoteUnitRelation`
+  - Added ``Windows::Check-Membership``
+  - Added ``Windows::Convert-SIDToFriendlyName``
+  - Refactored ``Juju::Is-JujuMasterUnit``
+  - Refactored ``Juju::Set-JujuRelation``
+  - Refactored ``Juju::Get-JujuRelationParams``
+  - Refactored ``Windows::Unmarshall-Object``
+  - Refactored ``Windows::Marshall-Object``
+  - Refactored ``Windows::Get-JujuUnitName``
+  - Refactored ``Windows::Create-LocalAdmin``
+
 ###version 0.13
   - 10.05.2015
   - Added ``Windows::Remove-CharmState``